% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bakdash_dist_measures}
\alias{bakdash_dist_measures}
\title{x}
\format{
A data frame with 230 rows 7 columns
\tabular{lll}{
\code{[, 1]} \tab \code{Subject}        \tab {Unique identifer for each participant} \cr
\code{[, 2]} \tab \code{Physical Distance} \tab {x} \cr
\code{[, 3]} \tab \code{Blindwalk Away}        \tab {x} \cr
\code{[, 4]} \tab \code{Blindwalk Toward} \tab {x} \cr
\code{[, 5]} \tab \code{Triangulated BW} \tab {x} \cr
\code{[, 6]} \tab \code{Verbal} \tab {x} \cr
\code{[, 7]} \tab \code{Visual Matching} \tab {x} \cr
}
}
\source{
x
}
\usage{
bakdash_dist_measures
}
\description{
x
}
