% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bland1995}
\alias{bland1995}
\title{Repeated measurements of intramural pH and PaCO2}
\format{
A data frame with 47 rows and 3 variables
\tabular{lll}{
\code{[, 1]} \tab \code{Subject} \tab {Unique identifer} \cr
\code{[, 2]} \tab \code{pH}      \tab {Potential of hydrogen, 
                                       acidity to base} \cr
\code{[, 3]} \tab \code{PaCO2}   \tab {Partial pressure of carbon dioxide}\cr     
}
}
\source{
Bland, J.M., & Altman, D.G. (1995). Calculating correlation
coefficients with repeated observations: Part 1 - correlation within
subjects. \emph{BMJ, 310}, 446, 
\doi{https://doi.org/10.1136/bmj.310.6977.446}
}
\usage{
bland1995
}
\description{
A dataset containing the repeated measurements of intramural pH
and PaCO2 for eight subjects, from Bland & Altman (1995).
}
