% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_rmc.R
\name{geom_rmc}
\alias{geom_rmc}
\title{geom_rmc: ggplot2 geom for simplified graphing}
\usage{
geom_rmc(my.rmc)
}
\arguments{
\item{my.rmc}{an object of class "rmc" generated from the \code{\link{rmcorr}} function.}
}
\description{
geom_rmc: ggplot2 geom for simplified graphing
}
