% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gilden2010}
\alias{gilden2010}
\title{Repeated measurements of reaction time and accuracy}
\format{
A data frame with 44 rows and 4 variables
\tabular{lll}{
\code{[, 1]} \tab \code{sub} \tab {Subject ID} \cr
\code{[, 2]} \tab \code{block} \tab {Block ID} \cr
\code{[, 3]} \tab \code{rt} \tab {Mean reaction time} \cr
\code{[, 4]} \tab \code{acc} \tab {Mean accuracy} \cr
}
}
\source{
Gilden, D.L., Thornton, T.L., & Marusich, L.R. (2010). The serial
process in visual search. \emph{Journal of Experimental Psychology: Human 
Perception and Performance, 36}, 533-542.
}
\usage{
gilden2010
}
\description{
A dataset containing four repeated measurements of reaction time
(RT) and accuracy from eleven subjects in a visual search experiment. Each
measurement is the mean RT and accuracy from a block of 288 search trials.
blocks of visual search, for eleven subjects
}
