% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.rmc.R
\name{glance.rmc}
\alias{glance.rmc}
\title{Glance for a(n) rmcorr object}
\usage{
\method{glance}{rmc}(x, ...)
}
\arguments{
\item{x}{A(n) rmcorr object created by rmcorr::rmcorr().}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A single row summary "glance" \code{\link[tibble:tibble]{tibble::tibble()}} of the rmc model.
\itemize{
\item \code{rho}: the value of the repeated measures correlation (rmcorr) coefficient (note rho is technically incorrect; this value is neither the population correlation nor a Spearman correlation. It was the closest match in the glossary)
\item \code{conf.low}: the lower-bound confidence interval for the rmcorr effect size
\item \code{conf.high}: the upper-bound confidence interval for the rmcorr effect size
\item \code{statistic}: the model F-statistic
\item \code{p.value}: the model p-value
\item \code{df.residual}: the model error/residual degrees of freedom (denominator)
\item \code{nobs}: the total number of paired observations
\item \code{n}: the number of unique participants (sample size)
}
}
\description{
Glance for a(n) rmcorr object
}
\examples{
my.rmc <- rmcorr(Subject, PaCO2, pH, bland1995)
glance(my.rmc)
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[=rmcorr]{rmcorr()}}
}
