% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raz2005}
\alias{raz2005}
\title{Repeated measurements of age and cerebellar volume}
\format{A data frame with 144 rows and 4 variables
\tabular{lll}{
\code{[, 1]} \tab \code{Participant} \tab {Participant ID} \cr
\code{[, 2]} \tab \code{Time} \tab {Measurement time} \cr
\code{[, 3]} \tab \code{Age} \tab {Participant's age (years)} \cr
\code{[, 4]} \tab \code{Volume} \tab {Adjusted volume of cerebellar 
hemispheres (cm^3)} \cr
}}
\source{
Raz, N., Lindenberger, U., Rodrigue, K.M., Kennedy, K.M., Head, D., 
Williamson, A., Dahle, C., Gerstorf, D., & Acker, J.D. (2005). Regional brain
 changes in aging healthy adults: General trends, individual differences, and
modifiers. \emph{Cerebral Cortex, 15}, 1676-1689.
}
\usage{
raz2005
}
\description{
A dataset containing two repeated measures, on two occasions
(\code{Time}), of age and adjusted volume of cerebellar hemispheres from 72
participants. Data were captured from Figure 8, Cerebellar Hemispheres
(lower right) of Raz et al. (2005).
}

