% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport_generics.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\alias{glance}
\alias{required_pkgs}
\title{Objects exported from other packages}
\seealso{
\code{\link[=tidy.rmc]{tidy.rmc()}}

\code{\link[=glance.rmc]{glance.rmc()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{required_pkgs}}, \code{\link[generics]{tidy}}}
}}

