% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmcorr-package.R
\docType{package}
\name{rmcorr-package}
\alias{rmcorr-package}
\title{A package for computing the repeated measures correlation coefficient}
\description{
\packageDescription{rmcorr}
}
\details{
details
}

