% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.rmcorr.R
\name{tidy.rmcorr}
\alias{tidy.rmcorr}
\title{Tidy for a(n) rmcorr object}
\usage{
\method{tidy}{rmcorr}(x, ...)
}
\arguments{
\item{x}{An rmcorr object created by rmcorr::rmcorr().}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble::()] summarizing component-level 
information about the model.
}
\description{
Tidy for a(n) rmcorr object
}
\examples{
my.rmc <- rmcorr(Subject, PaCO2, pH, bland1995)
}
\references{
Robinson D, Hayes A, Couch S, Hvitfeldt E (2025). 
\emph{broom: Convert Statistical Objects into Tidy Tibbles.} 
R package version 1.0.11, https://broom.tidymodels.org/.
}
\seealso{
[tidy()], [rmcorr::rmcorr()]
}
