% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{twedt_dist_measures}
\alias{twedt_dist_measures}
\title{Repeated measures and multivariate measures of perceived distance}
\format{
A data frame with 230 rows 7 columns
\tabular{lll}{
\code{[, 1]} \tab \code{Subject}           \tab {Unique identifier for each 
                                                 participant} \cr
\code{[, 2]} \tab \code{Physical Distance} \tab {Physical distance from the 
                                                 participant to the target 
                                                 cone, in meters} \cr
\code{[, 3]} \tab \code{Blindwalk Away}    \tab {Participants put on the 
                                                 blindfold after viewing the 
                                                 target. Next, participants 
                                                 took one step to the left 
                                                 and turned 180 degrees to 
                                                 face the opposite direction. 
                                                 Participants were instructed 
                                                 to walk forward until they 
                                                 had walked the original 
                                                 distance to the target.} \cr
\code{[, 4]} \tab \code{Blindwalk Toward}  \tab {Participants put on the 
                                                 blindfold after viewing the 
                                                 target. Next, participants 
                                                 walked forward until they 
                                                 thought they had reached the 
                                                 target cone. } \cr
\code{[, 5]} \tab \code{Triangulated BW}   \tab {Triangulated Blindwalk: 
                                                 Participants put on the 
                                                 blindfold after viewing the 
                                                 target. Next, participants
                                                 turned right 90 degrees 
                                                 and walked forward 5 meters. 
                                                 The experimenter told 
                                                 participants when to stop 
                                                 walking. Finally, 
                                                 participants turned to face 
                                                 toward the target and walked 
                                                 forward two steps.} \cr
\code{[, 6]} \tab \code{Verbal}            \tab {Participants stated the 
                                                 distance between the target 
                                                 cone and themselves, 
                                                 in feet and inches} \cr
\code{[, 7]} \tab \code{Visual Matching}   \tab {An experimenter stood next to
                                                 the target cone and walked 
                                                 away from the cone in a 
                                                 straight line that was 
                                                 perpendicular to the extent 
                                                 between the target and the 
                                                 participant. Participants 
                                                 instructed the experimenter 
                                                 to stop walking when they 
                                                 thought that the distance 
                                                 between the target and the 
                                                 experimenter was equal to 
                                                 the the target distance.} \cr
}
}
\source{
Dataset: Add Zenodo link w/doi? Unpublished manuscript.
}
\usage{
twedt_dist_measures
}
\description{
A dataset of repeated measures of distance perception  
at  physical distances of 7, 8, 9, 10, and 11 meters. The data are also 
multivariate five  dependent measures of distance perception. This is a 
5 (physical distance) x 5 (dependent measure) within-participants design 
with a sample size of 46. 
Note data is missing for 15 trials due participant and experimenter errors.
}
