% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bland1995}
\alias{bland1995}
\title{Repeated measurements of intramural pH and PaCO2}
\format{A data frame with 47 rows and 3 variables
\itemize{
\item \code{[, 1] Subject}
\item \code{[, 2] pH}
\item \code{[, 3] PacO2}
}}
\source{
Bland, J.M., & Altman, D.G. (1995). Calculating correlation
coefficients with repeated observations: Part 1 - correlation within
subjects. \emph{BMJ, 310}, 446.
}
\usage{
bland1995
}
\description{
A dataset containing the repeated measurements of intramural pH
and PaCO2 for eight subjects, from Bland & Altman (1995).
}

