% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glance.rmc.R
\name{glance.rmc}
\alias{glance.rmc}
\title{Glance for a(n) rmcorr object}
\usage{
\method{glance}{rmc}(x, ...)
}
\arguments{
\item{x}{A(n) rmcorr object created by rmcorr::rmcorr().}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A single row summary "glance" \code{\link[tibble:tibble]{tibble::tibble()}} of the rmc model.
}
\description{
Glance for a(n) rmcorr object
}
\examples{
my.rmc <- rmcorr(Subject, PaCO2, pH, bland1995)
glance(my.rmc)
}
\seealso{
\code{\link[=glance]{glance()}}, \code{\link[=rmcorr]{rmcorr()}}
}
