% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcorr.R
\name{print.rmc}
\alias{print.rmc}
\title{Calculate the repeated measures correlation coefficient.}
\usage{
\method{print}{rmc}(x, ...)
}
\arguments{
\item{x}{An object of class "rmc", a result of a call to rmcorr.}

\item{...}{additional arguments to \code{\link[base]{print}}.}
}
\description{
Calculate the repeated measures correlation coefficient.
}
\examples{
## Bland Altman 1995 data
blandrmc <- rmcorr(Subject, PacO2, pH, bland1995)
blandrmc
}
\seealso{
\code{\link{rmcorr}}
}
