% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.rmcorr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{tidy}
\title{Tidy for a(n) rmcorr object}
\arguments{
\item{x}{rmcorr output}
}
\examples{
my.rmc <- rmcorr(Subject, PaCO2, pH, bland1995)
tidy.rmcorr(my.rmc)
}
\references{
Robinson D, Hayes A, Couch S, Hvitfeldt E (2025). 
\emph{broom: Convert Statistical Objects into Tidy Tibbles.} 

R package version 1.0.11, https://broom.tidymodels.org/.
usethis::use_package("generics", "Imports")
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{tidy}}}
}}

