% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmcorr-package.R
\docType{package}
\name{rmcorr-package}
\alias{rmcorr-package}
\title{A package for computing the repeated measures correlation coefficient}
\description{
\packageDescription{rmcorr}
}
\references{
Bakdash, J.Z. & Marusich, L.R. (2017). 
Repeated Measures Correlation, \emph{Frontiers in Psychology, 8}, 456, 
\doi{https://doi.org/10.3389/fpsyg.2017.00456}

Bakdash, J.Z. & Marusich, L.R. (2019). 
Corrigendum: Repeated Measures Correlation, \emph{Frontiers in Psychology, 
10}, \doi{https://doi.org/10.3389/fpsyg.2019.01201}
}
