% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmcorr.R
\name{rmcorr}
\alias{rmcorr}
\title{Calculate the repeated measures correlation coefficient.}
\usage{
rmcorr(participant, measure1, measure2, dataset)
}
\arguments{
\item{participant}{A variable giving the subject name/id for each observation.}

\item{measure1}{A numeric variable giving the observations for one measure.}

\item{measure2}{A numeric variable giving the observations for the second measure.}

\item{dataset}{The data frame containing the variables.}
}
\value{
A list with class "rmc" containing the following components.
\item{r}{the value of the repeated measures correlation coefficient.}
\item{CI}{the 95\% confidence interval for the repeated measures correlation coefficient.}
\item{p}{the p-value for the repeated measures correlation coefficient.}
}
\description{
Calculate the repeated measures correlation coefficient.
}
\examples{
## Bland Altman 1995 data
rmcorr(Subject, pH, PacO2, bland1995)
}
\seealso{
\code{\link{rmcplot}}
}

