% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmcplot.R
\name{rmcplot}
\alias{rmcplot}
\title{Plot the repeated measures correlation coefficient.}
\usage{
rmcplot(participant, measure1, measure2, dataset, overall = F, xlab = NULL,
  ylab = NULL, ...)
}
\arguments{
\item{participant}{a variable giving the subject name/id for each observation.}

\item{measure1}{a numeric variable giving the observations for one measure.}

\item{measure2}{a numeric variable giving the observations for the second measure.}

\item{dataset}{the data frame containing the variables.}

\item{overall}{logical: if TRUE, plots the regression line between measure1 and
measure2, ignoring the participant variable.}

\item{xlab}{label for the x axis, defaults to the variable name for measure1.}

\item{ylab}{label for the y axis, defaults to the variable name for measure2.
names for measure1 and measure2.}

\item{...}{additional arguments to \code{\link[graphics]{plot}}.}
}
\description{
\code{rmcplot}  produces a scatterplot of \code{measure1} on the x-axis and
\code{measure2} on the y-axis, with a different color used for each subject.
Parallel lines are fitted to each subject's data.
}
\examples{
## Bland Altman 1995 data
rmcplot(participant = Subject, measure1 = PacO2, measure2 = pH,
dataset = bland1995, overall = TRUE)
}
\seealso{
\code{\link{rmcorr}}
}

