% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy.rmc.R
\name{tidy.rmc}
\alias{tidy.rmc}
\title{Tidy for a(n) rmcorr object}
\usage{
\method{tidy}{rmc}(x, ...)
}
\arguments{
\item{x}{A(n) 'rmcorr' object created by rmcorr::rmcorr().}

\item{...}{Unused, included for generic consistency only.}
}
\value{
A tidy [tibble::tibble::()] summarizing component-level 
information about the model.
}
\description{
Tidy for a(n) rmcorr object
}
\examples{
my.rmc <- rmcorr(Subject, PaCO2, pH, bland1995)
tidy(my.rmc)
}
\seealso{
[tidy()], [rmcorr::rmcorr()]
}
